+++ timestamps.php
@@ -8,13 +8,13 @@
  * that starts the plugin.
  *
  * @link              https://www.scoredetect.com/
- * @since             1.7.0
+ * @since             1.8.0
  * @package           SDCOM_Timestamps
  *
  * @wordpress-plugin
  * Plugin Name:       Timestamps
  * Description:       Timestamp your WordPress content to empower your content authenticity and increase user trust. No blockchain skills needed.
- * Version:           1.7.0
+ * Version:           1.8.0
  * Author:            ScoreDetect.com
  * Author URI:        https://www.scoredetect.com/
  * License:           AGPL-3.0-only
@@ -25,13 +25,15 @@
 
 namespace SDCOM_Timestamps;
 
+use WP_CLI;
+
 // If this file is called directly, abort.
 if ( ! defined( 'WPINC' ) ) {
 	die;
 }
 
 // Useful global constants.
-define( 'SDCOM_TIMESTAMPS_VERSION', '1.7.0' );
+define( 'SDCOM_TIMESTAMPS_VERSION', '1.8.0' );
 define( 'SDCOM_TIMESTAMPS_OPTIONS', 'sdcom_timestamps' );
 define( 'SDCOM_TIMESTAMPS_URL', plugin_dir_url( __FILE__ ) );
 define( 'SDCOM_TIMESTAMPS_PATH', plugin_dir_path( __FILE__ ) );
@@ -118,6 +120,13 @@ require_once SDCOM_TIMESTAMPS_INC . 'blocks.php';
  * Setup screen.
  */
 Screen::factory();
+
+/**
+ * WP CLI Commands.
+ */
+if ( defined( 'WP_CLI' ) && WP_CLI ) {
+	WP_CLI::add_command( 'timestamps', __NAMESPACE__ . '\Command' );
+}
 
 /**
  * Load text domain.